+++ python-client-sdk/src/sdk/models/operations/nullableoneoftypeinobjectpost.py
@@ -22,10 +22,10 @@ class NullableOneOfTypeInObjectPostRes:
 class NullableOneOfTypeInObjectPostResponse:
     content_type: str = dataclasses.field()
     r"""HTTP response content type for this operation"""
+    raw_response: requests_http.Response = dataclasses.field()
+    r"""Raw HTTP response; suitable for custom response parsing"""
     status_code: int = dataclasses.field()
     r"""HTTP response status code for this operation"""
-    raw_response: Optional[requests_http.Response] = dataclasses.field(default=None)
-    r"""Raw HTTP response; suitable for custom response parsing"""
     res: Optional[NullableOneOfTypeInObjectPostRes] = dataclasses.field(default=None)
     r"""OK"""