+++ typescript-client-sdk/src/sdk/models/operations/apikeyauthglobalnew.ts
@@ -23,5 +23,5 @@ export class ApiKeyAuthGlobalNewResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 }