+++ typescript-client-sdk/src/sdk/models/operations/oauth2auth.ts
@@ -41,7 +41,7 @@ export class Oauth2AuthResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * Successful authentication.