+++ typescript-client-sdk/src/sdk/models/operations/responsebodyxmlget.ts
@@ -22,7 +22,7 @@ export class ResponseBodyXmlGetResponse extends SpeakeasyBase {
      * Raw HTTP response; suitable for custom response parsing
      */
     @SpeakeasyMetadata()
-    rawResponse?: AxiosResponse;
+    rawResponse: AxiosResponse;
 
     /**
      * OK